removes dependency testing, adds ArrayableTrait
